requirements.txt
Flask==3.0.3
requests==2.31.0
beautifulsoup4==4.12.3
python-dateutil==2.9.0.post0
google-genai==0.4.0
tavily-python==0.3.3
flask-cors==4.0.0
python-dotenv==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
bcrypt==4.1.2
python-dotenv==1.0.1
